Adjust View and Mock Data to take in multiple sub-sessions
